Move navbar under navigation/
